Move runCommand into util.go.
