+++ docker/plugin_test.go
@@ -4,11 +4,6 @@
 package docker
 
 import (
-	"fmt"
-	"os"
-	"os/exec"
-	"testing"
-
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 	"gopkg.in/juju/charm.v5"
@@ -151,128 +146,24 @@ func (pluginSuite) TestDestroy(c *gc.C) {
 }
 
 type call struct {
-	out    []byte
+	out []byte
+	err error
+
 	argsIn []string
-	err    error
 }
 
 func fakeRunCommand(calls ...call) func([]string) ([]byte, error) {
 	index := 0
 	return func(args []string) ([]byte, error) {
+		calls[index].argsIn = args
 		call := calls[index]
 		index += 1
 
-		call.argsIn = args
 		if call.err != nil {
 			return nil, call.err
 		}
 		return call.out, nil
 	}
-}
-
-// fakeExecCommand replaces the normal exec.Command call to produce executables.
-// It returns a command that calls this test executable, telling it to run our
-// TestExecHelper test.  The original command and arguments are passed as
-// arguments to the testhelper after a "--" argument.
-func fakeExecCommand(name string, args ...string) *exec.Cmd {
-	args = append([]string{"-test.run=TestExecHelper", "--", name}, args...)
-	cmd := exec.Command(os.Args[0], args...)
-	cmd.Env = []string{"GO_WANT_HELPER_PROCESS=1"}
-	return cmd
-}
-
-func fakeExecCommandError(name string, args ...string) *exec.Cmd {
-	cmd := fakeExecCommand(name, args...)
-	cmd.Env = append(cmd.Env, "GO_HELPER_PROCESS_ERROR=1")
-	return cmd
-}
-
-// TestExecHelper is a fake test that is just used to do predictable things when
-// we run commands.
-func TestExecHelper(*testing.T) {
-	if os.Getenv("GO_WANT_HELPER_PROCESS") != "1" {
-		return
-	}
-	args := getTestArgs()
-	shouldErr := os.Getenv("GO_HELPER_PROCESS_ERROR") == "1"
-	if shouldErr {
-		defer os.Exit(1)
-	} else {
-		defer os.Exit(0)
-	}
-	switch args[0] {
-	case "run":
-		doRun(shouldErr)
-	case "inspect":
-		doInspect(shouldErr)
-	case "stop":
-		doStop(shouldErr)
-	case "rm":
-		doRm(shouldErr)
-	}
-}
-
-func doStop(shouldErr bool) {
-	if shouldErr {
-		fmt.Fprintln(os.Stderr, "Error response from daemon: no such id: foo")
-		return
-	}
-	// when docker successfully stops a container, it prints out the container id.
-	fmt.Fprintln(os.Stdout, "somebigid")
-}
-
-func doRm(shouldErr bool) {
-	if shouldErr {
-		fmt.Fprintln(os.Stderr, "Error response from daemon: no such id: foo")
-		return
-	}
-	// when docker successfully removes a container, it prints out the container id.
-	fmt.Fprintln(os.Stdout, "somebigid")
-}
-
-func doRun(shouldErr bool) {
-	if shouldErr {
-		fmt.Fprintf(os.Stderr, `Unable to find image 'foo:latest' locally
-Pulling repository foo
-Error: image library/foo:latest not found
-`)
-		return
-	}
-	fmt.Fprintln(os.Stdout, "somebigid")
-}
-
-func doInspect(shouldErr bool) {
-	if shouldErr {
-		fmt.Fprintln(os.Stderr, "Error: No such image or container: wegwgwg")
-		fmt.Fprintln(os.Stdout, "[]")
-		return
-	}
-	fmt.Fprint(os.Stdout, fakeInspectOutput)
-}
-
-// getTestArgs returns the args being passed to docker, so arg[0] would be the
-// docker command, like "run" or "stop".  This function will exit out of the
-// helper exec if it was not passed at least 3 arguments (e.g. docker stop id),
-// or if the first arg is not "docker".
-func getTestArgs() []string {
-	args := os.Args
-	for len(args) > 0 {
-		if args[0] == "--" {
-			args = args[1:]
-			break
-		}
-		args = args[1:]
-	}
-	if len(args) < 3 {
-		fmt.Fprintf(os.Stderr, "Not enough arguments passed to docker: %#v\n", args)
-		os.Exit(2)
-	}
-	if args[0] != "docker" {
-		fmt.Fprintf(os.Stderr, "Calling %q instead of docker\n", args[0])
-		os.Exit(2)
-	}
-
-	return args[1:]
 }
 
 const fakeInspectOutput = `
